// src/types/payload-types.ts
export interface Media {
  id: string;
  alt: string;
  caption?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
// ...
export interface Certificate {
  id: string;
  title: string;
  organization: string;
  certificateImage?: (string | null) | Media;
  certificateFile?: (string | null) | Media;
  issueDate?: string | null;
  credentialUrl?: string | null;
  order?: number | null;
  updatedAt: string;
  createdAt: string;
}